Generic JSON config reader
